% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The main application server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}
}
\description{
The main application server
}
